Expose bson stuff visible in the shell
Previusly only the `ObjectId` function was exposed and available
directly from the root of the `mongojs` module.

This commit exposes a couple of other BSON types in the same manner.
This is now the full list of exposed types:

- `mongojs.ObjectId`
- `mongojs.DBRef`
- `mongojs.Timestamp`
- `mongojs.MinKey`
- `mongojs.MaxKey`
- `mongojs.NumberLong` (the BSON type `Long`)

This is actually a revival commit, bringing back the spirit of the very
first Pull Request ever made to this projeect back in 2011 (later
refined in commit 06bbf02):

https://github.com/mafintosh/mongojs/pull/1